Rename the folder adding 'tab_' when appropriate
