+++ ggg_frontend/src/components/download/test-backend-interface.js
@@ -1,5 +1,4 @@
 import axios from 'axios';
-//import config from './config';
 
 export function postExample(url, data) {
   console.log(url);
