+++ ggg_frontend/src/tab_library/plaquette.js
